# documents/apidoc.yml
swagger: "2.0"
info:
  title: "Travel Tunes API"
  description: "API documentation for the Travel Tunes application"
  version: "1.0.0"
  contact:
    name: Nghi Quyen
    email: nghi.quyen@tuni.fi
basePath: "/api/v1"
...
paths:
...
  /users:
    get:
      tags:
        - "Users"
      summary: "Get all users"
      security:
        - BearerAuth: []   # Specify authentication requirement
      responses:
        200:
          description: "Success"
    post:
      tags:
        - "Users"
      summary: "Create a new user"
      responses:
        201:
          description: "User created"
...
  /reviews:
    get:
      tags:
        - "Reviews"
      summary: "Get all reviews"
      responses:
        200:
          description: "Success"
    post:
      tags:
        - "Reviews"
      summary: "Create a new review"
      security:
        - BearerAuth: []   # Specify authentication requirement
        - UserRole: ['user', 'admin']  # Specify user roles requirement
      responses:
        201:
          description: "Review created"
...